moved functionality of activity oncreate functions to abstract class
