+++ app/src/main/res/layout/activity_hybrids.xml
@@ -12,7 +12,7 @@
         android:layout_height="wrap_content" />
 
     <Button
-        android:id="@+id/hySuspend"
+        android:id="@+id/suspend"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/suspend_button"
@@ -20,20 +20,20 @@
         />
 
     <Button
-        android:id="@+id/hyPrev"
+        android:id="@+id/prev"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/previous_button"
-        android:layout_below="@id/hySuspend"
+        android:layout_below="@id/suspend"
         android:layout_alignParentLeft="true"
         />
 
     <Button
-        android:id="@+id/hyNext"
+        android:id="@+id/next"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/next_button"
-        android:layout_below="@id/hySuspend"
+        android:layout_below="@id/suspend"
         android:layout_alignParentRight="true"
         />
